refactoring

+++ resources/views/label/index.blade.php
@@ -1,41 +1,51 @@
 <x-app-layout>
     <h1 class="mb-5">{{ __('main.labels') }}</h1>
 
-    <div>
-        <a href="{{ route('labels.create') }}" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">
-            {{ __('label.create') }}
-        </a>
-    </div>
+    @can('create', App\Models\Label::class)
+        <div>
+            <a href="{{ route('labels.create') }}" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">
+                {{ __('label.create') }}
+            </a>
+        </div>
+    @endcan
 
-    <table class="mt-4">
-        <thead class="border-b-2 border-solid border-black text-left">
-            <tr>
-                <th>{{ __('label.id') }}</th>
-                <th>{{ __('label.name') }}</th>
-                <th>{{ __('label.description') }}</th>
-                <th>{{ __('label.created_at') }}</th>
-                <th>{{ __('table.actions') }}</th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach ($labels as $label)
-            <tr class="border-b border-dashed text-left">
-                <td>{{ $label->id }}</td>
-                <td>{{ $label->name }}</td>
-                <td>{{ $label->description }}</td>
-                <td>{{ $label->created_at }}</td>
-                <td>
-                    <a data-confirm="{{ __('table.confirm') }}"
-                    data-method="delete"
-                    class="text-red-600 hover:text-red-900"
-                    href="{{ route('labels.destroy', $label->id) }}">{{ __('table.delete') }}</a>
-                    <a class="text-blue-600 hover:text-blue-900"
-                    href="{{ route('labels.edit', $label->id) }}">{{ __('table.edit') }}</a>
-                </td>
-            </tr>
-            @endforeach
-        </tbody>
-    </table>
+    <x-table>
+        <x-slot name="thead">
+            <x-table.tr>
+                <x-table.th class="text-center">{{ __('label.id') }}</x-table.th>
+                <x-table.th>{{ __('label.name') }}</x-table.th>
+                <x-table.th>{{ __('label.description') }}</x-table.th>
+                <x-table.th>{{ __('label.created_at') }}</x-table.th>
+                @auth
+                    <x-table.th class="text-center">{{ __('table.actions') }}</x-table.th>
+                @endauth
+            </x-table.tr>
+        </x-slot>
+
+        @foreach ($labels as $label)
+            <x-table.tr>
+                <x-table.td class="text-center">{{ $label->id }}</x-table.td>
+                <x-table.td>{{ $label->name }}</x-table.td>
+                <x-table.td>{{ $label->description }}</x-table.td>
+                <x-table.td>{{ $label->created_at }}</x-table.td>
+                @auth
+                    <x-table.td class="space-x-3 text-center">
+                        @can('delete', $label)
+                            <a data-confirm="{{ __('table.confirm') }}"
+                                data-method="delete"
+                                class="text-red-600 hover:text-red-900"
+                                href="{{ route('labels.destroy', $label->id) }}">{{ __('table.delete') }}</a>
+                        @endcan
+
+                        @can('update', $label)
+                            <a class="font-medium text-blue-600 dark:text-blue-500 hover:underline"
+                                href="{{ route('labels.edit', $label->id) }}">{{ __('table.edit') }}</a>
+                        @endcan
+                    </x-table.td>
+                @endauth
+            </x-table.tr>
+        @endforeach
+    </x-table>
 
     <div class="mt-4">
         {{ $labels->links() }}
